'Done-Parse-Two'

+++ ServiceOTP/src/main/java/com/example/demo/service/OTPService.java
@@ -2,6 +2,7 @@ package com.example.demo.service;
 
 import java.text.ParseException;
 import java.time.LocalDateTime;
+import java.util.Date;
 import java.util.List;
 import java.util.Random;
 
@@ -21,7 +22,7 @@ public class OTPService {
 	private JavaMailSender mailSender;
 
 	public String generateOTP(String email) throws ParseException {
-
+		System.out.println(email);
 		Random random = new Random();
 
 		int otpValue = 100000 + random.nextInt(900000);
@@ -29,29 +30,37 @@ public class OTPService {
 		LocalDateTime expiry = now.plusMinutes(1);
 
 		OTP otp = new OTP(email, String.valueOf(otpValue), now, expiry);
+		try {
+			otpRepository.save(otp);
 
-		otpRepository.save(otp);
-
-		sendOTPByEmail(email, otp.getMessage());
-		
-		return otp.getMessage();
+			sendOTPByEmail(email, otp.getMessage());
+		} catch (Exception e) {
+			// TODO: handle exception
+			return "OTP has been sent. Please try again after 5 minutes.";
+		}
+	return "Send OTP Success";
 	}
 
 	public String checkingOTP(String email, String otp) {
-		OTP o = otpRepository.findOTP(email);
-		String rs = "fail";
-		if (o.getMessage().equalsIgnoreCase(otp)) {
-			rs = "success";
-			otpRepository.delete(o);
+		OTP o = null;
+		o = otpRepository.findOTP(email);
+		if(o != null) {
+			String rs = "OTP Invalid";
+			if (o.getMessage().equalsIgnoreCase(otp)) {
+				rs = "OTP Valid";
+				otpRepository.delete(o);
+			}
+			return rs;
 		}
-		return rs;
+		return "OTP Expired";
 	}
 
 	public void sendOTPByEmail(String toEmail, String otp) throws ParseException {
 		SimpleMailMessage message = new SimpleMailMessage();
+
 		message.setTo(toEmail);
 		message.setSubject("Xác thực OTP");
-		message.setText("Mã OTP của bạn là " + otp +". Mã OTP sẽ hết hạn sau 5 phút nữa");
+		message.setText(new Date().toString() + "Mã OTP của bạn là " + otp + ". Mã OTP sẽ hết hạn sau 5 phút nữa");
 		message.setFrom("gaming@gmail.com");
 
 		mailSender.send(message);
